requirements.txt
anthropic>=0.40.0
chainlit>=1.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.44.0
python-dotenv>=1.0.0
